sound/soc/aml/m8: Use devm_kfree instead of kfree
On the factory tester the sound chips are disabled in the device tree.
This means that the driver probe fails.  The problem is that the error
path in the probe function and the remove function of this driver are
both buggy and call kfree on a pointer that was allocated using
devm_kzalloc, resulting in heap corruption and a variety of panics
later on during the boot process.  Switching these functions to the
correct devm_kfree function fixes the panics.

Signed-off-by: Fionn Cleary <fionn.cleary@streamunlimited.com>